Abstract finalize and tests for process
